requirements.txt
discord.py>=2.3.0
asyncio
Pillow>=9.0.0